Slice out sdk message interface.
