+++ lib/rtm-client.js
@@ -3,6 +3,7 @@
 const EventEmitter = require('events');
 const mqtt = require('mqtt');
 const promised = require('./responses/promised');
+const getSdkMessage = require('./responses/message');
 
 class RtmClient extends EventEmitter {
   constructor(
@@ -98,45 +99,3 @@ class RtmClient extends EventEmitter {
 }
 
 module.exports = RtmClient;
-
-/**
- * Generalize real-time messages.
- *
- * @param channel
- * @param buffer
- * @return {{channel: *, data: *, raw: Buffer}}
- */
-function getSdkMessage({channel, buffer}) {
-  let data;
-  let decoded = false;
-  let error = null;
-
-  function decode() {
-    try {
-      data = JSON.parse(buffer.toString());
-    } catch (err) {
-      error = err;
-    }
-  }
-
-  return {
-    channel,
-    get data() {
-      if (decoded === false) {
-        decode();
-        decoded = true;
-      }
-
-      return data;
-    },
-    get error() {
-      if (decoded === false) {
-        decode();
-        decoded = true;
-      }
-
-      return error;
-    },
-    raw: buffer,
-  };
-}
